==> 02-data_cleanning.ipynb <==
import re
import numpy as np
import pandas as pd
from joblib import dump, load
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import SnowballStemmer

from text_cleaning import clean_up

categories = [
    'faciales',
    'capilares',
    'corporales',
    'maquillaje',
    'fragancias',
    'higiene'
]

dictionary = {
    each_category:pd.read_json(
        f'./data/{each_category}.json', orient='records'
    ) for each_category in categories
}

for each_category in categories:
    dictionary[each_category]['product_category'] = each_category

dictionary['faciales'].head()

df = pd.concat(dictionary.values())
df.head()

df.isnull().sum(axis=0)

df.product_category.value_counts()

df.product_chunk.str.split(expand=True).head()

df[['product_code', 'product_price']] = df.product_chunk.str.split(expand=True)[[1,5]]

df.head()

df['product_price'] = df.product_price.str.replace(r'\$', '').astype(int)

df.drop(columns=['product_chunk'], inplace=True)

df.head()

df.drop_duplicates(subset='product_code', inplace=True)

df.info()

df = df[[
    'product_code', 'product_name', 'product_price',
    'product_category', 'product_description',
    'product_more_info', 'product_image',
    
]]
df.head()

df.to_json('./data/products.json', orient='records', indent=2)

# create a column with all words from text columns
text_cols = ['product_name', 'product_category', 'product_description', 'product_more_info']

text_series = df[text_cols].apply(lambda x: " ".join(x), axis=1).apply(clean_up)
text_series.head()

dump(text_series, './data/text_series.joblib')
