==> 03-recommender_system.ipynb <==
# NLP
# import nltk
# nltk.download()

import pandas as pd
from joblib import load

from myroutines import clean_up, get_recommendation

# load products
df = pd.read_json('./data/products.json', orient='records')
df.head()

# load clean text
text_series = load('./data/text_series.joblib')
text_series.head()

# Construct a reverse map of indices and product codes
# indices = pd.Series(df.index, index=df.product_code).drop_duplicates()

my_description = '''
Cabello limpio, liso y radiante 
'''

recommendation = get_recommendation(my_description, text_series, df)
recommendation

recommendation.to_json('recommendation.json', orient='records')
